\name{@ANNOBJPREFIX@MAPCOUNTS} 
\alias{@ANNOBJPREFIX@MAPCOUNTS}
\title{Quality control information for @ANNOBJPREFIX@}
\description{
   @ANNOBJPREFIX@MAPCOUNTS is an R object that provides quality control
   information for @ANNOBJPREFIX@
}
\details{
  This file contains quality control information that can be displayed
  by typing @ANNOBJPREFIX@() after loading the package using
  library(@ANNOBJPREFIX@). The follow items are included: 

  Date built: Date when the package was built.

  Mappings found: total number of mappings obtained
}
\keyword{datasets}
